fix: table header z index and allowance tooltip z index
